refactor: Modularize utility functions
